requirements.txt
alabaster==0.7.12
altgraph==0.17.3
arrow==1.2.3
astroid==2.14.2
asttokens==2.0.5
atomicwrites==1.4.0
attrs==22.1.0
autopep8==1.6.0
Babel==2.11.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==0.0
bleach==4.1.0
Brotli==1.0.9
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
cookiecutter==1.7.3
cryptography==41.0.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.7
docstring-to-markdown==0.11
docutils==0.18.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==0.8.3
fastjsonschema==2.16.2
flake8==6.0.0
h11==0.14.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.0.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.25.0
ipython==8.12.2
ipython-genutils==0.2.0
isort==5.9.3
jaraco.classes==3.2.1
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyterlab-pygments==0.1.2
keyboard==0.13.5
keyring==23.13.1
Lackey==0.7.3
lazy-object-proxy==1.6.0
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==0.8.4
more-itertools==8.12.0
mouse==0.7.1
mypy-extensions==0.4.3
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
numpy==1.25.2
numpydoc==1.5.0
opencv-python==4.8.0.76
outcome==1.2.0
packaging==23.1
pandocfilters==1.5.0
paramiko==2.8.1
parso==0.8.3
pathspec==0.10.3
pefile==2023.2.7
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip==23.2.1
platformdirs==3.10.0
pluggy==1.0.0
ply==3.11
poyo==0.5.0
prompt-toolkit==3.0.36
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.0.1
PyGetWindow==0.0.9
Pygments==2.15.1
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
pylint==2.16.2
pylint-venv==2.3.0
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtWebEngine==5.15.4
PyRect==0.2.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.2
python-slugify==5.0.2
pytoolconfig==1.2.5
pytz==2022.7
pywin32==305.1
pywin32-ctypes==0.2.2
PyYAML==6.0
pyzmq==25.1.0
QDarkStyle==3.0.2
qstylizer==0.2.2
QtAwesome==1.2.2
qtconsole==5.4.2
QtPy==2.2.0
requests==2.31.0
rope==1.7.0
Rtree==1.0.1
selenium==4.12.0
setuptools==68.0.0
sip==6.6.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.4.3
spyder-kernels==2.4.4
stack-data==0.2.0
text-unidecode==1.3
textdistance==4.2.1
three-merge==0.1.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
tornado==6.3.2
traitlets==5.7.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
ujson==5.4.0
Unidecode==1.2.0
urllib3==1.26.16
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
whatthepatch==1.0.2
wheel==0.38.4
win-inet-pton==1.1.0
wrapt==1.14.1
wsproto==1.2.0
yapf==0.31.0
zipp==3.11.0
